use db from data.sql

+++ src/main/java/th/kmitl/Traveler/model/Placerepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Placerepository extends JpaRepository<Place, Integer> {
+public interface Placerepository extends JpaRepository<Place, Long> {
 
 }
